entertainment/requirements.txt
django>=5.1.7
python-decouple>=3.8
Pillow>=11.1.0
django-bootstrap5>=25.1
requests>=2.32.3
djangorestframework>=3.15.2
drf-spectacular>=0.28.0
django-q2>=1.7.6
redis>=5.2.1
psycopg>=3.2.6
gunicorn>=23.0.0
numpy>=2.2.4
whitenoise==6.9.0
psycopg2-binary>=2.9.10
django-debug-toolbar>=4.4.6
django-silk>=5.1.0
networkx>=3.2.1
pywebpush>=1.14.0